requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
Babel==2.10.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
chat-downloader==0.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.1.0
colorlog==6.6.0
dask==2022.6.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.3
django-annoying==0.10.6
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-compressor==4.0
django-libsass==0.9
django-phonenumber-field==6.3.0
django-select2==7.10.0
docstring-parser==0.14.1
et-xmlfile==1.1.0
filelock==3.6.0
fsspec==2022.5.0
idna==3.3
isodate==0.6.1
kombu==5.2.4
libsass==0.21.0
locket==1.0.0
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
partd==1.2.0
phonenumberslite==8.12.50
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.1
PyYAML==6.0
rcssmin==1.1.0
reportlab==3.6.9
requests==2.27.1
rjsmin==1.2.0
sentry-sdk==1.31.0
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
urllib3==1.26.11
uWSGI==2.0.20
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.0
xlrd==2.0.1
XlsxWriter==3.0.3
